# docs/api/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "AURA Care-Chain API",
        "description": "API REST do AURA Care-Chain \u2014 assistente de sa\u00fade domiciliar voice-first\npara idosos com Parkinson, com cadeia log\u00edstica de seguran\u00e7a da casa.\n\nConsumida por tr\u00eas clientes: app Flutter (Android/iOS/Web),\napp React Native e painel administrativo Angular.\n\nRegras de ouro: nunca prescreve nem diagnostica, o escore \u00e9 sempre\nexplic\u00e1vel (fatores + pesos) e nenhum dado de sa\u00fade \u00e9 gravado sem\nconsentimento LGPD.\n",
        "contact": {
            "name": "Equipe AURA \u2014 FIAP Smart HAS"
        },
        "license": {
            "name": "Uso acad\u00eamico \u2014 FIAP 2026"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/catalog": {
            "get": {
                "tags": [
                    "4. Cat\u00e1logo"
                ],
                "summary": "Lista o cat\u00e1logo, opcionalmente filtrado pelo risco que o item mitiga",
                "operationId": "list_3",
                "parameters": [
                    {
                        "name": "riskTag",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CatalogItemResponse"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha inesperada \u2014 a mensagem interna fica no log do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "INTERNAL_ERROR": {
                                        "value": {
                                            "error": {
                                                "code": "INTERNAL_ERROR",
                                                "message": "Erro interno.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token ausente, inv\u00e1lido ou expirado \u2014 ou credenciais incorretas no login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "UNAUTHORIZED": {
                                        "value": {
                                            "error": {
                                                "code": "UNAUTHORIZED",
                                                "message": "Autentica\u00e7\u00e3o necess\u00e1ria.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "NOT_FOUND": {
                                        "value": {
                                            "error": {
                                                "code": "NOT_FOUND",
                                                "message": "Casa n\u00e3o encontrada(a).",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Papel sem permiss\u00e3o (RBAC) ou recurso de outro paciente (isolamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                },
                                "examples": {
                                    "FORBIDDEN": {
                                        "value": {
                                            "error": {
                                                "code": "FORBIDDEN",
                                                "message": "Acesso negado a este recurso.",
                                                "details": null
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CatalogItemResponse": {
                "type": "object",
                "properties": {
                    "sku": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "installable": {
                        "type": "boolean"
                    },
                    "normRef": {
                        "type": "string"
                    },
                    "riskTag": {
                        "type": "string"
                    },
                    "stockNearby": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ApiErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/Body"
                    }
                }
            },
            "Body": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "details": {
                        "type": "object"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /api/v1/auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}